+++ packages/ui/accordion/stories/Accordion.stories.tsx
@@ -40,7 +40,11 @@ const Accordion = ({
 						noSideBorders={noSideBorders}
 						isFlushed={isFlushed}
 					>
-						<AccordionItem title="Tab Title 1" contentShadow={false}>
+						<AccordionItem
+							title="Tab Title 1"
+							hasToggle={true}
+							contentShadow={false}
+						>
 							<AccordionItemBody>
 								<p>
 									Lorem Ipsum is simply dummy text of the printing and