+++ packages/ui/table/CHANGELOG.md
@@ -1,5 +1,34 @@
 # Change Log
 
+## 1.0.6
+
+### Patch Changes
+
+- [#355](https://github.com/wpmudev/sui-react/pull/355)
+  [`503754f`](https://github.com/wpmudev/sui-react/commit/503754fe4f0d7745369493f06311bbc55d9322f6)
+  Thanks [@creador-dev](https://github.com/creador-dev)! - Popover popup
+  automatic placement.
+
+- [#355](https://github.com/wpmudev/sui-react/pull/355)
+  [`c37ddcb`](https://github.com/wpmudev/sui-react/commit/c37ddcb96d23a3af499de507cec2100b75264c4e)
+  Thanks [@creador-dev](https://github.com/creador-dev)! - Setup banner and
+  navigation custom icon prop
+
+- Updated dependencies
+  [[`503754f`](https://github.com/wpmudev/sui-react/commit/503754fe4f0d7745369493f06311bbc55d9322f6),
+  [`c37ddcb`](https://github.com/wpmudev/sui-react/commit/c37ddcb96d23a3af499de507cec2100b75264c4e),
+  [`36ba29c`](https://github.com/wpmudev/sui-react/commit/36ba29c920888612cf431d0a419cbf8a42e9838e)]:
+  - @wpmudev/sui-form-field@0.1.1
+  - @wpmudev/sui-icons@0.0.3
+  - @wpmudev/sui-checkbox@0.0.6
+  - @wpmudev/sui-dropdown@0.2.1
+  - @wpmudev/sui-button@0.1.3
+  - @wpmudev/sui-select@0.2.1
+  - @wpmudev/sui-input@0.1.1
+  - @wpmudev/sui-grid@0.1.3
+  - @wpmudev/sui-box@0.1.3
+  - @wpmudev/sui-hooks@0.1.2
+
 ## 1.0.5
 
 ### Patch Changes
@@ -8,6 +37,10 @@
   [`0a9e7dd`](https://github.com/wpmudev/sui-react/commit/0a9e7dda13f63ed410df711027ef032d05a41654)
   Thanks [@creador-dev](https://github.com/creador-dev)! - Update colors
 
+- [#355](https://github.com/wpmudev/sui-react/pull/355)
+  [`ef804ad`](https://github.com/wpmudev/sui-react/commit/ef804adbfb5d783e066897e4771bc201f704351b)
+  Thanks [@creador-dev](https://github.com/creador-dev)! - Remove sui react
+  package.
 - [#334](https://github.com/wpmudev/sui-react/pull/334)
   [`de4695f`](https://github.com/wpmudev/sui-react/commit/de4695feb4886a3af66e5c4b51031915907e2490)
   Thanks [@abder](https://github.com/abder)! - Remove sui react package.
@@ -17,6 +50,8 @@
   Thanks [@creador-dev](https://github.com/creador-dev)! - Add new template card
   component.
 
+- [#355](https://github.com/wpmudev/sui-react/pull/355)
+  [`ef804ad`](https://github.com/wpmudev/sui-react/commit/ef804adbfb5d783e066897e4771bc201f704351b)
 - [#361](https://github.com/wpmudev/sui-react/pull/361)
   [`a9ead37`](https://github.com/wpmudev/sui-react/commit/a9ead374fd02ebd63882d55bc01ab2fc7375732f)
   Thanks [@creador-dev](https://github.com/creador-dev)! - Change ids to
@@ -29,6 +64,16 @@
 
 - Updated dependencies
   [[`0a9e7dd`](https://github.com/wpmudev/sui-react/commit/0a9e7dda13f63ed410df711027ef032d05a41654),
+  [`ef804ad`](https://github.com/wpmudev/sui-react/commit/ef804adbfb5d783e066897e4771bc201f704351b),
+  [`ef804ad`](https://github.com/wpmudev/sui-react/commit/ef804adbfb5d783e066897e4771bc201f704351b),
+  [`4a48b89`](https://github.com/wpmudev/sui-react/commit/4a48b899de749dd8e73169dd0096f293e557b0a1),
+  [`ef804ad`](https://github.com/wpmudev/sui-react/commit/ef804adbfb5d783e066897e4771bc201f704351b)]:
+  - @wpmudev/sui-checkbox@0.0.5
+  - @wpmudev/sui-form-field@0.1.0
+  - @wpmudev/sui-dropdown@0.2.0
+  - @wpmudev/sui-select@0.2.0
+  - @wpmudev/sui-input@0.1.0
+  - @wpmudev/sui-button@0.1.2
   [`de4695f`](https://github.com/wpmudev/sui-react/commit/de4695feb4886a3af66e5c4b51031915907e2490),
   [`4a48b89`](https://github.com/wpmudev/sui-react/commit/4a48b899de749dd8e73169dd0096f293e557b0a1),
   [`a9ead37`](https://github.com/wpmudev/sui-react/commit/a9ead374fd02ebd63882d55bc01ab2fc7375732f)]: